LangChain_Faiss/pyproject.toml
[project]
name = "langchain-faiss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "asgiref==3.9.1",
    "attrs==25.3.0",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "colorama==0.4.6",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "django-redis==6.0.0",
    "faiss-cpu==1.12.0",
    "flask>=3.1.2",
    "frozenlist==1.7.0",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "idna==3.10",
    "jiter==0.10.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.27",
    "langchain-community==0.3.29",
    "langchain-core==0.3.75",
    "langchain-openai==0.3.32",
    "langchain-text-splitters==0.3.10",
    "langsmith==0.4.21",
    "marshmallow==3.26.1",
    "multidict==6.6.4",
    "mypy-extensions==1.1.0",
    "numpy==2.3.2",
    "openai==1.102.0",
    "orjson==3.11.3",
    "packaging==25.0",
    "propcache==0.3.2",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "python-dotenv==1.1.1",
    "pyyaml==6.0.2",
    "redis==6.4.0",
    "regex==2025.8.29",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.43",
    "sqlparse==0.5.3",
    "tenacity==9.1.2",
    "tiktoken==0.11.0",
    "tqdm==4.67.1",
    "typing-extensions==4.14.1",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
    "yarl==1.20.1",
    "zstandard==0.24.0",
]
